requirements.txt
absl-py
altair
annotated-types
astunparse
attrs
blinker
blis
cachetools
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
colorama
confection
contourpy
cycler
cymem
decorator
flatbuffers
fonttools
gast
gitdb
GitPython
google-pasta
grpcio
idna
imageio
imageio-ffmpeg
jax
jaxlib
Jinja2
jsonschema
jsonschema-specifications
keras
kiwisolver
langcodes
language_data
libclang
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
mediapipe
moviepy
murmurhash
namex
narwhals
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
opt-einsum
optree
packaging
pandas
pillow
preshed
proglog
protobuf
pyarrow
pycparser
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
pytz
referencing
requests
rich
rpds-py
scipy
setuptools
shellingham
six
smart-open
smmap
sounddevice
spacy
spacy-legacy
spacy-loggers
srsly
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-intel
termcolor
thinc
toml
tornado
tqdm
typer
typing_extensions
tzdata
urllib3
wasabi
watchdog
weasel
Werkzeug
wheel
wrapt
opencv-python-headless
numpy
pillow
pip
streamlit
tensorflow==2.16.2
opencv-python-headless
numpy
pillow
pip==24.2

